Change JS manifest to application.js
Not sure if the naming was intentional to prevent overwriting existing files, but when the generator is run it creates a 'applications.js' file instead of an 'application.js' file. After running the generator, two JS manifest files are created which causes a asset pipeline error since one of them automatically includes the other.