Bump less-rails to remove deprecation warnings, fixes #925
The warning is:

```
DEPRECATION WARNING: Sprockets method `register_engine` is deprecated.
Please register a mime type using `register_mime_type` then
use `register_compressor` or `register_transformer`.
```

This problem is really a less-rails error, and is documented in this
issue of that repo:

https://github.com/metaskills/less-rails/issues/122

It was fixed with this PR, merged Oct 2017:

https://github.com/metaskills/less-rails/pull/137

Bumping our less-rails dependency to 3.0.0 will give us the updates in
that PR.

Version 3.0.0 is the next version after 2.8.0, so there are very few
other changes to worry about.  You can view all of the less-rails code
changes between 2.8.0 and 3.0.0 here:

https://github.com/metaskills/less-rails/compare/05186eb...7bb14e5

They are almost all test and/or README changes. So this update should be
low risk.